Correction sur choisir formation et suppression formation 80%

+++ backend/microserviceGIR/src/main/java/sn/ugb/gir/service/impl/InscriptionAdministrativeFormationServiceImpl.java
@@ -1,7 +1,6 @@
 package sn.ugb.gir.service.impl;
 
 import java.time.Instant;
-import java.time.LocalDate;
 import java.util.Objects;
 import java.util.Optional;
 import org.slf4j.Logger;
@@ -10,8 +9,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import sn.ugb.gir.domain.InscriptionAdministrative;
 import sn.ugb.gir.domain.InscriptionAdministrativeFormation;
+import sn.ugb.gir.domain.PaiementFrais;
 import sn.ugb.gir.repository.InscriptionAdministrativeFormationRepository;
+import sn.ugb.gir.repository.InscriptionAdministrativeRepository;
+import sn.ugb.gir.repository.PaiementFraisRepository;
 import sn.ugb.gir.repository.search.InscriptionAdministrativeFormationSearchRepository;
 import sn.ugb.gir.service.InscriptionAdministrativeFormationService;
 import sn.ugb.gir.service.dto.*;
@@ -34,16 +37,22 @@ public class InscriptionAdministrativeFormationServiceImpl implements Inscriptio
 
     private final InscriptionAdministrativeFormationRepository inscriptionAdministrativeFormationRepository;
 
+    private final InscriptionAdministrativeRepository inscriptionAdministrativeRepository;
+
+    private final PaiementFraisRepository paiementFraisRepository;
+
     private final InscriptionAdministrativeFormationMapper inscriptionAdministrativeFormationMapper;
 
     private final InscriptionAdministrativeFormationSearchRepository inscriptionAdministrativeFormationSearchRepository;
 
     public InscriptionAdministrativeFormationServiceImpl(
-        InscriptionAdministrativeFormationRepository inscriptionAdministrativeFormationRepository,
-        InscriptionAdministrativeFormationMapper inscriptionAdministrativeFormationMapper,
-        InscriptionAdministrativeFormationSearchRepository inscriptionAdministrativeFormationSearchRepository
+            InscriptionAdministrativeFormationRepository inscriptionAdministrativeFormationRepository, InscriptionAdministrativeRepository inscriptionAdministrativeRepository, PaiementFraisRepository paiementFraisRepository,
+            InscriptionAdministrativeFormationMapper inscriptionAdministrativeFormationMapper,
+            InscriptionAdministrativeFormationSearchRepository inscriptionAdministrativeFormationSearchRepository
     ) {
         this.inscriptionAdministrativeFormationRepository = inscriptionAdministrativeFormationRepository;
+        this.inscriptionAdministrativeRepository = inscriptionAdministrativeRepository;
+        this.paiementFraisRepository = paiementFraisRepository;
         this.inscriptionAdministrativeFormationMapper = inscriptionAdministrativeFormationMapper;
         this.inscriptionAdministrativeFormationSearchRepository = inscriptionAdministrativeFormationSearchRepository;
     }
@@ -54,8 +63,9 @@ public class InscriptionAdministrativeFormationServiceImpl implements Inscriptio
         InscriptionAdministrativeFormation inscriptionAdministrativeFormation = inscriptionAdministrativeFormationMapper.toEntity(
             inscriptionAdministrativeFormationDTO
         );
-        Instant currentDate = Instant.from(LocalDate.now());
+        Instant currentDate = Instant.now();
 
+        validateData( inscriptionAdministrativeFormationDTO);
         inscriptionAdministrativeFormationDTO.setDateChoixFormation(currentDate);
         inscriptionAdministrativeFormation = inscriptionAdministrativeFormationRepository.save(inscriptionAdministrativeFormation);
         InscriptionAdministrativeFormationDTO result = inscriptionAdministrativeFormationMapper.toDto(inscriptionAdministrativeFormation);
@@ -69,6 +79,8 @@ public class InscriptionAdministrativeFormationServiceImpl implements Inscriptio
         InscriptionAdministrativeFormation inscriptionAdministrativeFormation = inscriptionAdministrativeFormationMapper.toEntity(
             inscriptionAdministrativeFormationDTO
         );
+
+        validateData( inscriptionAdministrativeFormationDTO);
         inscriptionAdministrativeFormation = inscriptionAdministrativeFormationRepository.save(inscriptionAdministrativeFormation);
         InscriptionAdministrativeFormationDTO result = inscriptionAdministrativeFormationMapper.toDto(inscriptionAdministrativeFormation);
         inscriptionAdministrativeFormationSearchRepository.index(inscriptionAdministrativeFormation);
@@ -81,6 +93,7 @@ public class InscriptionAdministrativeFormationServiceImpl implements Inscriptio
     ) {
         log.debug("Request to partially update InscriptionAdministrativeFormation : {}", inscriptionAdministrativeFormationDTO);
 
+        validateData( inscriptionAdministrativeFormationDTO);
         return inscriptionAdministrativeFormationRepository
             .findById(inscriptionAdministrativeFormationDTO.getId())
             .map(existingInscriptionAdministrativeFormation -> {
@@ -116,6 +129,11 @@ public class InscriptionAdministrativeFormationServiceImpl implements Inscriptio
     @Override
     public void delete(Long id) {
         log.debug("Request to delete InscriptionAdministrativeFormation : {}", id);
+
+        PaiementFrais paiementFrais = paiementFraisRepository.paiementFraisIAF(id);
+        if (paiementFrais.getEcheancePayeeYN()){
+            throw new BadRequestAlertException("Vous ne pouvez pas supprimer un paiement déja valider par le paiement", ENTITY_NAME, "InscriptionFormationDejaPayer");
+        }
         inscriptionAdministrativeFormationRepository.deleteById(id);
         inscriptionAdministrativeFormationSearchRepository.deleteFromIndexById(id);
     }
@@ -129,41 +147,42 @@ public class InscriptionAdministrativeFormationServiceImpl implements Inscriptio
             .map(inscriptionAdministrativeFormationMapper::toDto);
     }
 
-    public void validateData(InscriptionAdministrativeFormationDTO inscriptionAdministrativeFormationDTO, Long id ){
+    public void validateData(InscriptionAdministrativeFormationDTO inscriptionAdministrativeFormationDTO){
 
 
         if (Objects.equals(inscriptionAdministrativeFormationDTO.getFormation(),null)) {
             throw new BadRequestAlertException("Veuillez renseigné la formation à laquelle s'inscrit l'étudiant", "Formation", "FormationObligatoire");
         }
         if (Objects.equals(inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getEtudiant(),null)) {
-            throw new BadRequestAlertException("Veuillez renseigné l'etudiant qui doit s'inscrire ", "Etudiant", "EtudiantObligatoire");
+            throw new BadRequestAlertException("Veuillez renseigné l'etudiant qui doit s'inscrire ", ENTITY_NAME, "EtudiantObligatoire");
         }
         if (Objects.equals(inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getAnneeAcademique(),null)) {
-            throw new BadRequestAlertException("Veuillez renseigné l'année de l'inscription", "AnneeAcademique", "AnneeAcademiqueObligatoire");
+            throw new BadRequestAlertException("Veuillez renseigné l'année de l'inscription", ENTITY_NAME, "AnneeAcademiqueObligatoire");
         }
         if (Objects.equals(inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getTypeAdmission(),null)) {
-            throw new BadRequestAlertException("Veuillez renseigné le type d'admission de l'étudiant", "TypeAdmission", "TypeAdmissionObligatoire");
+            throw new BadRequestAlertException("Veuillez renseigné le type d'admission de l'étudiant", ENTITY_NAME, "TypeAdmissionObligatoire");
         }
+        inscriptionAdministrativeFormationDTO.setInscriptionPrincipaleYN(TRUE);
         Long formationId = inscriptionAdministrativeFormationDTO.getFormation().getId() ;
         Long etudiantId = inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getEtudiant().getId();
         Long AnneeAcademiqueId = inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getAnneeAcademique().getId();
         Optional<InscriptionAdministrativeFormation> existingIAF= inscriptionAdministrativeFormationRepository.findByFormationIdAndInscriptionAdministrativeEtudiantIdAndInscriptionAdministrativeAnneeAcademiqueId(formationId,etudiantId,AnneeAcademiqueId);
-        Optional<InscriptionAdministrativeFormation> existingIA= inscriptionAdministrativeFormationRepository.findByInscriptionAdministrativeEtudiantIdAndInscriptionAdministrativeAnneeAcademiqueId(etudiantId,AnneeAcademiqueId);
+        Page<InscriptionAdministrative> existingIA= inscriptionAdministrativeRepository.findByEtudiantIdAndAnneeAcademiqueId(etudiantId,AnneeAcademiqueId);
+        Long count = existingIA.getTotalElements();
+
         if (existingIAF.isPresent()) {
             if (!Objects.equals(existingIAF.get().getId(), inscriptionAdministrativeFormationDTO.getId())) {
-                throw new BadRequestAlertException("l'inscription de cet etudiant à cette formation est deja faite pour cet annee", "TypeAdmission", "TypeAdmissionObligatoire");
+                throw new BadRequestAlertException("l'inscription de cet etudiant à cette formation est deja faite pour cet annee", "InscriptionAdministrativeFormation", "InscriptionAdministrativeFormationExiste");
             }
         }
-        if (existingIA.isPresent()) {
-            if (Objects.equals(existingIA.get().getInscriptionAdministrative().getId(), inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getId())  ) {
-                if (Objects.equals( inscriptionAdministrativeFormationDTO.getFormation(), null)){
-                    inscriptionAdministrativeFormationDTO.setInscriptionPrincipaleYN(TRUE);
-                } else if ( !Objects.equals(existingIA.get().getFormation().getId(), inscriptionAdministrativeFormationDTO.getFormation().getId())) {
-                    inscriptionAdministrativeFormationDTO.setInscriptionPrincipaleYN(FALSE);
-                }
-
-            }
-        }
+//        if (existingIA.isPresent()) {
+//            if (Objects.equals(existingIA.get().getId(), inscriptionAdministrativeFormationDTO.getInscriptionAdministrative().getId())  ) {
+//               if ( !Objects.equals(existingIA.get().getFormation().getId(), inscriptionAdministrativeFormationDTO.getFormation().getId())) {
+//                    inscriptionAdministrativeFormationDTO.setInscriptionPrincipaleYN(FALSE);
+//                }
+//
+//            }
+//        }
 
     }
